refactor: 아이스크림 출력 관련 수정 Step2 아이스크림 그리기 부분 관련 수정 icecream.swift파일 생성 main.swift에 작성 한 icecream 그리기 관련 메서드를 iceream.swift에 옮김. 아이스크림 그리기 부분 주석처리.
