+++ CodeStarterCamp_Week1/main.swift
@@ -8,22 +8,7 @@
 
 import Foundation
 
-let hight = 12
-let width = 11
-let halfWidth = (width/2)+1
-
-func printIceCream() {
-    print("*", terminator: "")
-}
-
-func printStick(_ widthCount: Int) {
-    if widthCount == halfWidth-1 || widthCount == halfWidth+1 {
-        print("|", terminator: "")
-    } else {
-        print(" ", terminator: "")
-    }
-}
-
+/* 아이스크림 출력 부분
 for hightCount in 1...hight {
     for widthCount in 1...width {
         if hightCount < 9 {
@@ -34,7 +19,8 @@ for hightCount in 1...hight {
     }
     print("")
 }
-print("")
+*/
+
 printPeperoOnOrder(bodyLength: 10, body: "***", topping: "", stickLength: 4)
 printPeperoOnOrder(bodyLength: 12, body: "***", topping: "&", stickLength: 4)
 printPeperoOnOrder(bodyLength: 12, body: "***", topping: "#", stickLength: 6)
